TINY-8313: Moved the hr plugin functionality to core
