--- modules/bridge/src/demo/ts/plugins/HrItems.ts
@@ -1,10 +1,0 @@
-import { getDemoRegistry } from '../buttons/DemoRegistry';
-
-export const registerHrItems = (): void => {
-  getDemoRegistry().addButton('hr', {
-    type: 'button',
-    onAction: (_buttonApi) => {
-      // add hr to content
-    }
-  });
-};
