+++ modules/oxide/src/demo/index.html
@@ -41,7 +41,7 @@
 
         tinymce.init({
             selector:'textarea',
-            plugins: 'autoresize advlist anchor autolink charmap code codesample directionality fullscreen help hr image insertdatetime link lists media nonbreaking pagebreak preview searchreplace table template textpattern toc visualblocks visualchars wordcount',
+            plugins: 'autoresize advlist anchor autolink charmap code codesample directionality fullscreen help image insertdatetime link lists media nonbreaking pagebreak preview searchreplace table template textpattern toc visualblocks visualchars wordcount',
             toolbar: 'undo redo | formatselect | bold italic strikethrough forecolor backcolor blockquote | link table image media | alignleft aligncenter alignright alignjustify | numlist bullist outdent indent | removeformat fullscreen',
             toolbar_mode: 'floating',
             toolbar_sticky: true,
