+++ modules/tinymce/src/core/demo/ts/demo/CommandsDemo.ts
@@ -100,7 +100,7 @@ export default () => {
     skin_url: '../../../../js/tinymce/skins/ui/oxide',
     selector: 'textarea.tinymce',
     plugins: [
-      'advlist autolink link image lists charmap preview hr anchor pagebreak',
+      'advlist autolink link image lists charmap preview anchor pagebreak',
       'searchreplace wordcount visualblocks visualchars code fullscreen insertdatetime media nonbreaking',
       'save table directionality emoticons template paste importcss textpattern codesample'
     ],
