+++ modules/tinymce/src/core/demo/ts/demo/ContentSecurityPolicyDemo.ts
@@ -83,7 +83,7 @@ const settings = {
     makeSidebar(ed, 'sidebar1', 'green', 200);
   },
   plugins: [
-    'autosave advlist autolink link image lists charmap preview hr anchor pagebreak',
+    'autosave advlist autolink link image lists charmap preview anchor pagebreak',
     'searchreplace wordcount visualblocks visualchars code fullscreen insertdatetime media nonbreaking',
     'save table directionality emoticons template paste importcss textpattern codesample help noneditable'
   ],
