+++ modules/tinymce/src/core/demo/ts/demo/InlineDemo.ts
@@ -24,7 +24,7 @@ export default () => {
       { title: 'Some title 2', description: 'Some desc 2', content: '<div class="mceTmpl"><span class="cdate">cdate</span><span class="mdate">mdate</span>My content2</div>' }
     ],
     plugins: [
-      'autosave advlist autolink link image lists charmap preview hr anchor pagebreak',
+      'autosave advlist autolink link image lists charmap preview anchor pagebreak',
       'searchreplace wordcount visualblocks visualchars code fullscreen insertdatetime media nonbreaking',
       'save table directionality emoticons template paste textpattern',
       'codesample help noneditable'
