+++ modules/tinymce/src/core/demo/ts/demo/ResponsiveDemo.ts
@@ -78,7 +78,7 @@ export default () => {
     },
     plugins: [
       'fullscreen help',
-      'autosave advlist autolink link image lists charmap preview hr anchor pagebreak',
+      'autosave advlist autolink link image lists charmap preview anchor pagebreak',
       'searchreplace wordcount visualblocks visualchars code fullscreen insertdatetime media nonbreaking',
       'save table directionality emoticons template paste importcss textpattern',
       'codesample help noneditable'
