+++ modules/tinymce/src/core/demo/ts/demo/SourceDumpDemo.ts
@@ -11,7 +11,7 @@ export default () => {
     ],
     image_caption: true,
     plugins: [
-      'autosave advlist autolink link image lists charmap preview hr anchor pagebreak',
+      'autosave advlist autolink link image lists charmap preview anchor pagebreak',
       'searchreplace wordcount visualblocks visualchars code fullscreen insertdatetime media nonbreaking',
       'save table directionality emoticons template paste importcss textpattern',
       'codesample help noneditable'
