--- modules/tinymce/src/plugins/hr/demo/html/demo.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>Plugin: hr Demo Page</title>
-  </head>
-
-  <body>
-  <h2>Plugin: hr Demo Page</h2>
-    <div id="ephox-ui">
-      <textarea cols="30" rows="10" class="tinymce"></textarea>
-    </div>
-    <script src="../../../../../js/tinymce/tinymce.js"></script>
-    <script src="../../../../../scratch/demos/plugins/hr/demo.js"></script>
-  </body>
-</html>
