--- modules/tinymce/src/plugins/hr/demo/ts/demo/Demo.ts
@@ -1,10 +1,0 @@
-declare let tinymce: any;
-
-tinymce.init({
-  selector: 'textarea.tinymce',
-  plugins: 'hr code',
-  toolbar: 'hr code',
-  height: 600
-});
-
-export {};
