--- modules/tinymce/src/plugins/hr/main/ts/api/Commands.ts
@@ -1,18 +1,0 @@
-/**
- * Copyright (c) Tiny Technologies, Inc. All rights reserved.
- * Licensed under the LGPL or a commercial license.
- * For LGPL see License.txt in the project root for license information.
- * For commercial licenses see https://www.tiny.cloud/
- */
-
-import Editor from 'tinymce/core/api/Editor';
-
-const register = (editor: Editor): void => {
-  editor.addCommand('InsertHorizontalRule', () => {
-    editor.execCommand('mceInsertContent', false, '<hr />');
-  });
-};
-
-export {
-  register
-};
