--- modules/tinymce/src/plugins/hr/test/ts/browser/HrSanityTest.ts
@@ -1,27 +1,0 @@
-import { ApproxStructure } from '@ephox/agar';
-import { describe, it } from '@ephox/bedrock-client';
-import { TinyAssertions, TinyHooks, TinyUiActions } from '@ephox/wrap-mcagar';
-
-import Editor from 'tinymce/core/api/Editor';
-import Plugin from 'tinymce/plugins/hr/Plugin';
-
-describe('browser.tinymce.plugins.hr.HrSanitytest', () => {
-  const hook = TinyHooks.bddSetupLight<Editor>({
-    plugins: 'hr',
-    toolbar: 'hr',
-    base_url: '/project/tinymce/js/tinymce'
-  }, [ Plugin ]);
-
-  it('TBA: Click on the horizontal rule toolbar button and assert hr is added to the editor', () => {
-    const editor = hook.editor();
-    TinyUiActions.clickOnToolbar(editor, 'button[aria-label="Horizontal line"]');
-    TinyAssertions.assertContentStructure(editor, ApproxStructure.build((s) => {
-      return s.element('body', {
-        children: [
-          s.element('hr', {}),
-          s.anything()
-        ]
-      });
-    }));
-  });
-});
