+++ modules/tinymce/src/themes/silver/test/ts/browser/editor/menubar/EditorMenubarOptionsTest.ts
@@ -9,7 +9,7 @@ import Editor from 'tinymce/core/api/Editor';
 
 import { countNumber, extractOnlyOne } from '../../../module/UiUtils';
 
-describe('browser.tinymce.themes.silver.editor.menubar.EditorMenubarSettingsTest', () => {
+describe('browser.tinymce.themes.silver.editor.menubar.EditorMenubarOptionsTest', () => {
 
   const pCreateEditorWithMenubar = (menubar: boolean | string | undefined) => McEditor.pFromSettings<Editor>({
     menubar,
@@ -22,7 +22,7 @@ describe('browser.tinymce.themes.silver.editor.menubar.EditorMenubarSettingsTest
     'Checking structure of tox-menubar is "default"',
     ApproxStructure.build((s, str, arr) => s.element('div', {
       classes: [ arr.has('tox-menubar') ],
-      children: Arr.map([ 'File', 'Edit', 'View', 'Format' ], (x) =>
+      children: Arr.map([ 'File', 'Edit', 'View', 'Insert', 'Format' ], (x) =>
         s.element('button', {
           children: [
             s.element('span', {
